Fix data manager: add data_manager_conf.xml, move tool into data_manager/ subdir, install models via stanza.download (default_fast)
