+++ data_managers/data_manager_stanza_models/data_manager/data_manager_stanza_models.xml
@@ -1,7 +1,13 @@
-<tool id="data_manager_stanza_models" name="Stanza Language Models" version="1.11.1.3" tool_type="manage_data" profile="21.05">
+<tool id="data_manager_stanza_models" name="Stanza Language Models" version="@TOOL_VERSION@+galaxy@VERSION_SUFFIX@" tool_type="manage_data" profile="@PROFILE@">
     <description>Download and install Stanza language models</description>
+    <macros>
+        <token name="@TOOL_VERSION@">1.12.0</token>
+        <token name="@VERSION_SUFFIX@">0</token>
+        <token name="@PROFILE@">21.05</token>
+    </macros>
     <requirements>
         <requirement type="package" version="3.12">python</requirement>
+        <requirement type="package" version="@TOOL_VERSION@">stanza</requirement>
     </requirements>
     <command detect_errors="exit_code"><![CDATA[
         python '$__tool_directory__/data_manager_stanza_models.py'
@@ -90,9 +96,10 @@ the Stanza NLP annotation tool.
 How It Works
 ------------
 
-Stanza models are downloaded using the ``stanza.download()`` API and stored
-in a Galaxy-managed directory. Each language downloads all available processors
-(tokenization, POS, lemma, depparse, NER, etc.) as a single package.
+Stanza models are downloaded using the ``stanza.download()`` API and stored in a
+Galaxy-managed stanza_resources directory. Each language installs the memory-efficient
+``default_fast`` package (nocharlm models for tokenization, POS, lemma, depparse, NER,
+and sentiment where available), matching the models loaded by the Stanza NLP tool.
 
 Available Languages
 -------------------
@@ -114,7 +121,7 @@ multiple languages will take longer accordingly.
 Version
 -------
 
-This data manager installs models compatible with Stanza version 1.11.1.
+This data manager installs models compatible with Stanza version @TOOL_VERSION@.
 
 For more information about Stanza models, see:
 https://stanfordnlp.github.io/stanza/available_models.html
