+++ data_managers/data_manager_stanza_models/data_manager_conf.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<data_managers>
+    <data_manager tool_file="data_manager/data_manager_stanza_models.xml" id="data_manager_stanza_models">
+        <data_table name="stanza_models">
+            <output>
+                <column name="value" />
+                <column name="name" />
+                <column name="lang" />
+                <column name="models_path" />
+            </output>
+        </data_table>
+    </data_manager>
+</data_managers>
